feat: add basic logic for resolving relations
This commit introduces a breaking change for the "count" method. The
method now accepts an object -- like the "create" method does -- instead
of a mapper function. You can still specify such function via a special
"$transform" attribute.

	User.count(3, {
	  $transform: (usr, _) => {
	    usr.firstName = 'Loki';
	    return usr;
	  },
	});

Additionally, this commit introduces the special attributes which are
responsible for resolving model relations. Every attribute which starts
with either "$has" or "$for" will be treated as a special attribute.

	Post.create({ $forUser: 2 });
        Post.create({ $forUser: user }); // where user is an instance of User
        Post.create({ $forUser: true }); // autoresolve (not supported yet)

Note: I still need to figure out what happens if these special attributes conflict
with column names. I think that the probability of such conflicts is very low,
but I'll take it in consideration.